requirements.txt
python-telegram-bot==20.0
crunpyroll
ffmpeg-python==0.2.0
pymkv2
requests==2.28.1
aiohttp==3.8.3
cryptography==38.0.1
python-dotenv==0.21.0  # optionally Added for loading environment variables if isnt satisfied with config.py or if user is using .env in whole code instead of config.py 
